initial commit with phantom plugin refactor
