--- src/insilicoICH/ground_truth_definition/__init__.py
@@ -1,1 +1,0 @@
-from . import dicom_to_voxelized_phantom, phantoms, utils
