+++ src/insilicoICH/phantoms/__init__.py
@@ -1,0 +1,1 @@
+from . import base_phantoms, dicom_to_voxelized_phantom, utils
