+++ src/insilicoICH/study.py
@@ -17,11 +17,8 @@ from scipy.ndimage import center_of_mass
 from monai.transforms import RandAffine
 
 from .image_acquisition import Scanner, read_dicom
-from .ground_truth_definition.phantoms import (NIHPD_Head,
-                                               MIDA_Head,
-                                               Phantom,
-                                               possible_ages)
-from .ground_truth_definition import iq_phantoms
+from .phantoms.head_phantoms import MIDA_Head, NIHPD_Head, possible_ages
+# from .ground_truth_definition import iq_phantoms
 
 
 load_dotenv()
@@ -55,7 +52,8 @@ def load_vol(file_list):
     return np.stack(list(map(read_dicom, file_list)))
 
 
-available_phantoms = possible_ages + [o for o in dir(iq_phantoms) if (not o.startswith('__')) and o not in ['np', 'create_circle_phantom', 'Phantom', 'create_resolution_phantom', 'create_ct_phantom_with_bars']]
+available_phantoms = possible_ages
+# + [o for o in dir(iq_phantoms) if (not o.startswith('__')) and o not in ['np', 'create_circle_phantom', 'Phantom', 'create_resolution_phantom', 'create_ct_phantom_with_bars']]
 
 
 def load_phantom(name='Densitometry', shape=None):
