+++ tests/test_study.py
@@ -71,3 +71,5 @@ def test_subdural_augmented_position_study():
                     views=100)
     measured_lesion_signal = study.images[study.lesion.astype(bool)].mean()
     assert measured_lesion_signal > 21
+
+test_epidural_augmented_position_study()
